deal with unavailability of watchdog

+++ breakonthru/pico/main.py
@@ -85,7 +85,8 @@ class PicoDoorReceiver:
 
         #self.log(f'Managing state at time {self.now}')
 
-        self.watchdog.feed() # feed the watchdog timer to avoid board reboot
+        if self.watchdog is not None:
+            self.watchdog.feed() # feed the watchdog timer to avoid board reboot
 
         if self.unlocked:
             # the door is currently unlocked
